+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngineSession.kt
@@ -796,17 +796,6 @@ class GeckoEngineSession(
             }
         }
 
-        private fun recoverGeckoSession() {
-            // Recover the GeckoSession after the process getting killed or crashing. We create a
-            // new underlying GeckoSession.
-            // Eventually we may be able to re-use the same GeckoSession by re-opening it. However
-            // that seems to have caused issues:
-            // https://github.com/mozilla-mobile/android-components/issues/3640
-
-            geckoSession.close()
-            createGeckoSession()
-        }
-
         override fun onFullScreen(session: GeckoSession, fullScreen: Boolean) {
             notifyObservers { onFullScreenChange(fullScreen) }
         }