+++ mobile/android/android-components/components/browser/storage-sync/src/main/java/mozilla/components/browser/storage/sync/Types.kt
@@ -99,8 +99,6 @@ internal fun mozilla.appservices.places.uniffi.VisitTransition.into() = when (th
     mozilla.appservices.places.uniffi.VisitTransition.FRAMED_LINK -> VisitType.FRAMED_LINK
 }
 
-private val intToVisitType: Map<Int, VisitType> = VisitType.values().associateBy(VisitType::type)
-
 internal fun mozilla.appservices.places.uniffi.HistoryVisitInfo.into(): VisitInfo {
     return VisitInfo(
         url = this.url,