+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/DefaultComponents.kt
@@ -235,7 +235,7 @@ open class DefaultComponents(private val applicationContext: Context) {
 
     // Intent
     val tabIntentProcessor by lazy {
-        TabIntentProcessor(tabsUseCases, sessionUseCases.loadUrl, searchUseCases.newTabSearch)
+        TabIntentProcessor(tabsUseCases, searchUseCases.newTabSearch)
     }
     val externalAppIntentProcessors by lazy {
         listOf(